change name model userProfile => Profile
