+++ lib/app/modules/home/home_binding.dart
@@ -2,7 +2,7 @@ import 'package:get/get.dart';
 
 import 'package:habitat_ft_user/app/modules/home/home_controller.dart';
 
-import 'controllers/user_tile_controller.dart';
+import 'controllers/profile_controller.dart';
 import 'controllers/subscription_list_controller.dart';
 
 class HomeBinding extends Bindings {
@@ -12,8 +12,8 @@ class HomeBinding extends Bindings {
     Get.create<SubscriptionListController>(
       () => SubscriptionListController(),
     );
-    Get.lazyPut<UserTileController>(
-      () => UserTileController(),
+    Get.lazyPut<ProfileController>(
+      () => ProfileController(),
     );
   }
 }
