+++ lib/app/modules/home/widgets/profile.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:habitat_ft_user/app/utils/config/colors.dart';
 import 'package:habitat_ft_user/app/utils/config/styles.dart';
-import 'package:habitat_ft_user/app/modules/home/controllers/user_tile_controller.dart';
+import 'package:habitat_ft_user/app/modules/home/controllers/profile_controller.dart';
 
-class UserTile extends GetWidget<UserTileController> {
+class Profile extends GetWidget<ProfileController> {
   @override
   Widget build(BuildContext context) {
     return Padding(
