+++ lib/app/modules/root/root_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:habitat_ft_user/app/modules/home/services/user_profile_service.dart';
+import 'package:habitat_ft_user/app/modules/home/services/profile_service.dart';
 import 'package:habitat_ft_user/app/modules/home/services/subscription_service.dart';
 import 'package:habitat_ft_user/app/modules/login/services/auth_service.dart';
 
@@ -10,7 +10,7 @@ class RootBinding extends Bindings {
   void dependencies() {
     Get.put<RootController>(RootController());
     Get.put<AuthService>(AuthService());
-    Get.put<UserProfileService>(UserProfileService());
+    Get.put<ProfileService>(ProfileService());
     Get.put<SubscriptionService>(SubscriptionService());
   }
 }
